% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_sets.R
\name{add_ts_to_access_set}
\alias{add_ts_to_access_set}
\title{Add time series to access set}
\usage{
add_ts_to_access_set(access_set, ts_keys, ignore_missing = F)
}
\arguments{
\item{access_set}{Name of the access set. An access set is a group of time series for which a permission (read_quota, read, read_before_release, write) can be granted to individual users. Every time series can be a member of multiple access sets.}

\item{ts_keys}{Character vector of time series identifiers (keys)}

\item{ignore_missing}{Whether to ignore missing or forbidden time series when requesting time series data}
}
\description{
Adds existing time series (given by their keys) to a time series access set.
}
\seealso{
Other access set functions: 
\code{\link{create_access_set}()},
\code{\link{delete_access_set}()},
\code{\link{list_access_sets}()},
\code{\link{read_access_set_keys}()},
\code{\link{remove_ts_from_access_set}()}
}
\concept{access set functions}
