% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collections.R
\name{add_ts_to_collection}
\alias{add_ts_to_collection}
\title{Add time series to collection}
\usage{
add_ts_to_collection(collection, owner = "self", ts_keys, ignore_missing = F)
}
\arguments{
\item{collection}{Name of the time series collection}

\item{owner}{Username of the owner of the time series collection}

\item{ts_keys}{Character vector of time series identifiers (keys)}

\item{ignore_missing}{Whether to ignore missing or forbidden time series when requesting time series data}
}
\description{
Adds existing time series (given by their keys) to a time series collection.
}
\seealso{
Other time series collection functions: 
\code{\link{create_collection}()},
\code{\link{delete_collection}()},
\code{\link{list_collections}()},
\code{\link{read_collection_keys}()},
\code{\link{read_collection_ts}()},
\code{\link{read_collection_ts_history}()},
\code{\link{read_collection_ts_metadata}()},
\code{\link{read_collection_ts_release}()},
\code{\link{read_collection_ts_release_future}()},
\code{\link{read_collection_ts_release_history}()},
\code{\link{read_collection_ts_write_time}()},
\code{\link{remove_ts_from_collection}()}
}
\concept{time series collection functions}
