% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{add_user_access_sets}
\alias{add_user_access_sets}
\title{Add user access sets}
\usage{
add_user_access_sets(username = "self", access_sets, permission)
}
\arguments{
\item{username}{Username of the time series database user}

\item{access_sets}{Names of the access sets. An access set is a group of time series for which a permission (one of read_quota, read, read_before_release or write) can be granted to users. A time series can be a member of multiple access sets.}

\item{permission}{The permission granted to the user for the time series in the access set. Must be one of read_quota, read, read_before_release or write.}
}
\description{
Add access sets for a user with a given permission.
}
\seealso{
Other user management functions: 
\code{\link{create_user_api_key}()},
\code{\link{create_user_quota}()},
\code{\link{delete_user_quota}()},
\code{\link{list_user_access_sets}()},
\code{\link{list_users}()},
\code{\link{remove_user_access_sets}()},
\code{\link{write_user_quota}()}
}
\concept{user management functions}
