% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_sets.R
\name{create_access_set}
\alias{create_access_set}
\title{Create access set}
\usage{
create_access_set(access_set, description)
}
\arguments{
\item{access_set}{Name of the access set. An access set is a group of time series for which a permission (read_quota, read, read_before_release, write) can be granted to individual users. Every time series can be a member of multiple access sets.}

\item{description}{Description of the access set}
}
\description{
Create a new time series access set.
}
\seealso{
Other access set functions: 
\code{\link{add_ts_to_access_set}()},
\code{\link{delete_access_set}()},
\code{\link{list_access_sets}()},
\code{\link{read_access_set_keys}()},
\code{\link{remove_ts_from_access_set}()}
}
\concept{access set functions}
