% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collections.R
\name{create_collection}
\alias{create_collection}
\title{Create time series collection}
\usage{
create_collection(collection, owner = "self", description)
}
\arguments{
\item{collection}{Name of the time series collection}

\item{owner}{Username of the owner of the time series collection}

\item{description}{Description of the time series collection}
}
\description{
Create a new time series collection. By default, the created collection is owned by the authenticated user.
To create a collection for another user, provide the username of that user as the owner parameter.
}
\seealso{
Other time series collection functions: 
\code{\link{add_ts_to_collection}()},
\code{\link{delete_collection}()},
\code{\link{list_collections}()},
\code{\link{read_collection_keys}()},
\code{\link{read_collection_ts}()},
\code{\link{read_collection_ts_history}()},
\code{\link{read_collection_ts_metadata}()},
\code{\link{read_collection_ts_release}()},
\code{\link{read_collection_ts_release_future}()},
\code{\link{read_collection_ts_release_history}()},
\code{\link{read_collection_ts_write_time}()},
\code{\link{remove_ts_from_collection}()}
}
\concept{time series collection functions}
