% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{create_dataset}
\alias{create_dataset}
\title{Create a dataset}
\usage{
create_dataset(dataset, description)
}
\arguments{
\item{dataset}{Name of the dataset. A dataset is a group of time series with a common theme. Every time series can belong to only one dataset.}

\item{description}{Description of the dataset.}
}
\description{
Create a time series dataset. Every time series can belong to only one dataset.
}
\seealso{
Other dataset functions: 
\code{\link{delete_dataset}()},
\code{\link{list_datasets}()},
\code{\link{read_dataset_keys}()},
\code{\link{read_dataset_ts}()},
\code{\link{read_dataset_ts_history}()},
\code{\link{read_dataset_ts_metadata}()},
\code{\link{read_dataset_ts_release}()},
\code{\link{read_dataset_ts_release_future}()},
\code{\link{read_dataset_ts_release_history}()},
\code{\link{read_dataset_ts_write_time}()},
\code{\link{write_dataset_ts_release}()}
}
\concept{dataset functions}
