% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{create_user}
\alias{create_user}
\title{Create user}
\usage{
create_user(username, role)
}
\arguments{
\item{username}{Username of the time series database user}

\item{role}{Role of the user. Must be one of 'admin', 'intern' or 'extern'.}
}
\description{
Create a time series database user with a given role.
}
