% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{create_user_api_key}
\alias{create_user_api_key}
\title{Create user API key}
\usage{
create_user_api_key(username = "self")
}
\arguments{
\item{username}{Username of the time series database user}
}
\value{
List with API key.
}
\description{
Create an API key for a user for programmatic access.
Any previously created API key for the user will be overwritten and therefore invalidated.
Store the newly created API key securely because it cannot be retrieved later.
}
\seealso{
Other user management functions: 
\code{\link{add_user_access_sets}()},
\code{\link{create_user_quota}()},
\code{\link{delete_user_quota}()},
\code{\link{list_user_access_sets}()},
\code{\link{list_users}()},
\code{\link{remove_user_access_sets}()},
\code{\link{write_user_quota}()}
}
\concept{user management functions}
