% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{create_user_quota}
\alias{create_user_quota}
\title{Create user quota}
\usage{
create_user_quota(username, subscription_annual_quota, subscription_start_date)
}
\arguments{
\item{username}{Username of the time series database user}

\item{subscription_annual_quota}{Annual download quota of data service subscription}

\item{subscription_start_date}{Start date of the user subscription}
}
\description{
Create an annual quota for a data service subscriber. Without a quota, the user has an unlimited number of time series downloads.
}
\seealso{
Other user management functions: 
\code{\link{add_user_access_sets}()},
\code{\link{create_user_api_key}()},
\code{\link{delete_user_quota}()},
\code{\link{list_user_access_sets}()},
\code{\link{list_users}()},
\code{\link{remove_user_access_sets}()},
\code{\link{write_user_quota}()}
}
\concept{user management functions}
