% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{delete_dataset}
\alias{delete_dataset}
\title{Delete dataset}
\usage{
delete_dataset(dataset, skip_check = F)
}
\arguments{
\item{dataset}{Name of the dataset. A dataset is a group of time series with a common theme. Every time series can belong to only one dataset.}

\item{skip_check}{If FALSE (default), the user must manually confirm the deletion of time series.}
}
\description{
Permanently delete all time series in a dataset, including all vintages and metadata, and the dataset itself.
}
\seealso{
Other dataset functions: 
\code{\link{create_dataset}()},
\code{\link{list_datasets}()},
\code{\link{read_dataset_keys}()},
\code{\link{read_dataset_ts}()},
\code{\link{read_dataset_ts_history}()},
\code{\link{read_dataset_ts_metadata}()},
\code{\link{read_dataset_ts_release}()},
\code{\link{read_dataset_ts_release_future}()},
\code{\link{read_dataset_ts_release_history}()},
\code{\link{read_dataset_ts_write_time}()},
\code{\link{write_dataset_ts_release}()}
}
\concept{dataset functions}
