% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series.R
\name{delete_ts}
\alias{delete_ts}
\title{Delete time series}
\usage{
delete_ts(ts_keys, ignore_missing = F, skip_check = F)
}
\arguments{
\item{ts_keys}{Character vector of time series identifiers (keys)}

\item{ignore_missing}{Whether to ignore missing or forbidden time series when requesting time series data}

\item{skip_check}{If FALSE (default), the user must manually confirm the deletion of time series.}
}
\description{
Delete time series given by their unique identifiers (keys). The time series are removed completely, including all vintages and metadata.
}
\seealso{
Other time series functions: 
\code{\link{find_ts}()},
\code{\link{read_ts}()},
\code{\link{read_ts_dataset}()},
\code{\link{read_ts_history}()},
\code{\link{read_ts_metadata}()},
\code{\link{read_ts_release}()},
\code{\link{read_ts_release_future}()},
\code{\link{read_ts_release_history}()},
\code{\link{read_ts_write_time}()},
\code{\link{rename_ts}()},
\code{\link{write_ts}()},
\code{\link{write_ts_dataset}()},
\code{\link{write_ts_metadata}()},
\code{\link{write_ts_release}()}
}
\concept{time series functions}
