% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{delete_user_quota}
\alias{delete_user_quota}
\title{Delete user quota}
\usage{
delete_user_quota(username)
}
\arguments{
\item{username}{Username of the time series database user}
}
\description{
Delete the existing quota of a data service subscriber.
Without a quota, the user has an unlimited number of time series downloads.
}
\seealso{
Other user management functions: 
\code{\link{add_user_access_sets}()},
\code{\link{create_user_api_key}()},
\code{\link{create_user_quota}()},
\code{\link{list_user_access_sets}()},
\code{\link{list_users}()},
\code{\link{remove_user_access_sets}()},
\code{\link{write_user_quota}()}
}
\concept{user management functions}
