% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series.R
\name{find_ts}
\alias{find_ts}
\title{Find time series}
\usage{
find_ts(regexp)
}
\arguments{
\item{regexp}{Regular expression pattern to search for time series keys.}
}
\value{
Vector of time series keys matching the regular expression.
}
\description{
Find time series keys by regular expression.
}
\seealso{
Other time series functions: 
\code{\link{delete_ts}()},
\code{\link{read_ts}()},
\code{\link{read_ts_dataset}()},
\code{\link{read_ts_history}()},
\code{\link{read_ts_metadata}()},
\code{\link{read_ts_release}()},
\code{\link{read_ts_release_future}()},
\code{\link{read_ts_release_history}()},
\code{\link{read_ts_write_time}()},
\code{\link{rename_ts}()},
\code{\link{write_ts}()},
\code{\link{write_ts_dataset}()},
\code{\link{write_ts_metadata}()},
\code{\link{write_ts_release}()}
}
\concept{time series functions}
