% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_config}
\alias{get_config}
\title{Get package configuration}
\usage{
get_config()
}
\value{
Named list of options.
}
\description{
Get the package configuration options of the current session. Every option is initialized with the corresponding environment variables on package load.
}
\seealso{
Other package configuration functions: 
\code{\link{set_config}()}
}
\concept{package configuration functions}
