% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_sets.R
\name{list_access_sets}
\alias{list_access_sets}
\title{List access sets}
\usage{
list_access_sets()
}
\value{
Table with a row for every existing access set.
}
\description{
Read information on existing time series access sets.
}
\seealso{
Other access set functions: 
\code{\link{add_ts_to_access_set}()},
\code{\link{create_access_set}()},
\code{\link{delete_access_set}()},
\code{\link{read_access_set_keys}()},
\code{\link{remove_ts_from_access_set}()}
}
\concept{access set functions}
