% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{list_datasets}
\alias{list_datasets}
\title{List datasets}
\usage{
list_datasets()
}
\value{
Table with a row for every existing dataset.
}
\description{
Read information on existing time series datasets.
}
\seealso{
Other dataset functions: 
\code{\link{create_dataset}()},
\code{\link{delete_dataset}()},
\code{\link{read_dataset_keys}()},
\code{\link{read_dataset_ts}()},
\code{\link{read_dataset_ts_history}()},
\code{\link{read_dataset_ts_metadata}()},
\code{\link{read_dataset_ts_release}()},
\code{\link{read_dataset_ts_release_future}()},
\code{\link{read_dataset_ts_release_history}()},
\code{\link{read_dataset_ts_write_time}()},
\code{\link{write_dataset_ts_release}()}
}
\concept{dataset functions}
