% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{list_user_access_sets}
\alias{list_user_access_sets}
\title{List user access sets}
\usage{
list_user_access_sets(username = "self")
}
\arguments{
\item{username}{Username of the time series database user}
}
\value{
Table with a row for every user access set.
}
\description{
List the access sets and associated permissions of a particular user.
}
\seealso{
Other user management functions: 
\code{\link{add_user_access_sets}()},
\code{\link{create_user_api_key}()},
\code{\link{create_user_quota}()},
\code{\link{delete_user_quota}()},
\code{\link{list_users}()},
\code{\link{remove_user_access_sets}()},
\code{\link{write_user_quota}()}
}
\concept{user management functions}
