% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_sets.R
\name{read_access_set_keys}
\alias{read_access_set_keys}
\title{Read access set keys}
\usage{
read_access_set_keys(access_set)
}
\arguments{
\item{access_set}{Name of the access set. An access set is a group of time series for which a permission (read_quota, read, read_before_release, write) can be granted to individual users. Every time series can be a member of multiple access sets.}
}
\value{
Character vector of time series keys
}
\description{
Read the keys of the time series in a time series access set.
}
\seealso{
Other access set functions: 
\code{\link{add_ts_to_access_set}()},
\code{\link{create_access_set}()},
\code{\link{delete_access_set}()},
\code{\link{list_access_sets}()},
\code{\link{remove_ts_from_access_set}()}
}
\concept{access set functions}
