% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collections.R
\name{read_collection_keys}
\alias{read_collection_keys}
\title{Read collection time series keys}
\usage{
read_collection_keys(collection, owner = "self")
}
\arguments{
\item{collection}{Name of the time series collection}

\item{owner}{Username of the owner of the time series collection}
}
\value{
Character vector of time series keys.
}
\description{
Read the keys of the time series in a time series collection.
}
\seealso{
Other time series collection functions: 
\code{\link{add_ts_to_collection}()},
\code{\link{create_collection}()},
\code{\link{delete_collection}()},
\code{\link{list_collections}()},
\code{\link{read_collection_ts}()},
\code{\link{read_collection_ts_history}()},
\code{\link{read_collection_ts_metadata}()},
\code{\link{read_collection_ts_release}()},
\code{\link{read_collection_ts_release_future}()},
\code{\link{read_collection_ts_release_history}()},
\code{\link{read_collection_ts_write_time}()},
\code{\link{remove_ts_from_collection}()}
}
\concept{time series collection functions}
