% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collections.R
\name{read_collection_ts_metadata}
\alias{read_collection_ts_metadata}
\title{Read collection time series metadata}
\usage{
read_collection_ts_metadata(
  collection,
  owner = "self",
  locale = NULL,
  ignore_missing = F
)
}
\arguments{
\item{collection}{Name of the time series collection}

\item{owner}{Username of the owner of the time series collection}

\item{locale}{The locale of the metadata. Can be any string, but ISO codes are recommended (such as 'en', 'de', 'fr', 'it'). Set to NULL for unlocalized metadata (default).}

\item{ignore_missing}{Whether to ignore missing or forbidden time series when requesting time series data}
}
\value{
List of time series metadata. Each list element is named by the corresponding time series key and contains the metadata as a named list.
}
\description{
Read the time series metadata of a particular locale of the time series in a collection.
}
\seealso{
Other time series collection functions: 
\code{\link{add_ts_to_collection}()},
\code{\link{create_collection}()},
\code{\link{delete_collection}()},
\code{\link{list_collections}()},
\code{\link{read_collection_keys}()},
\code{\link{read_collection_ts}()},
\code{\link{read_collection_ts_history}()},
\code{\link{read_collection_ts_release}()},
\code{\link{read_collection_ts_release_future}()},
\code{\link{read_collection_ts_release_history}()},
\code{\link{read_collection_ts_write_time}()},
\code{\link{remove_ts_from_collection}()}
}
\concept{time series collection functions}
