% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collections.R
\name{read_collection_ts_release_future}
\alias{read_collection_ts_release_future}
\title{Read collection time series release future}
\usage{
read_collection_ts_release_future(
  collection,
  owner = "self",
  ignore_missing = F
)
}
\arguments{
\item{collection}{Name of the time series collection}

\item{owner}{Username of the owner of the time series collection}

\item{ignore_missing}{Whether to ignore missing or forbidden time series when requesting time series data}
}
\value{
Table with release topic, year, period and time for every time series key and future release.
}
\description{
Read the release information of future time series vintages.
}
\seealso{
Other time series collection functions: 
\code{\link{add_ts_to_collection}()},
\code{\link{create_collection}()},
\code{\link{delete_collection}()},
\code{\link{list_collections}()},
\code{\link{read_collection_keys}()},
\code{\link{read_collection_ts}()},
\code{\link{read_collection_ts_history}()},
\code{\link{read_collection_ts_metadata}()},
\code{\link{read_collection_ts_release}()},
\code{\link{read_collection_ts_release_history}()},
\code{\link{read_collection_ts_write_time}()},
\code{\link{remove_ts_from_collection}()}
}
\concept{time series collection functions}
