% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collections.R
\name{read_collection_ts_release_history}
\alias{read_collection_ts_release_history}
\title{Read collection time series release history}
\usage{
read_collection_ts_release_history(
  collection,
  owner = "self",
  valid_from = as.Date("1900-01-01"),
  valid_to = Sys.Date(),
  ignore_missing = F
)
}
\arguments{
\item{collection}{Name of the time series collection}

\item{owner}{Username of the owner of the time series collection}

\item{valid_from}{Start date of the vintage range}

\item{valid_to}{End date of the vintage range}

\item{ignore_missing}{Whether to ignore missing or forbidden time series when requesting time series data}
}
\value{
Table with vintage date, release topic, year, period and time for every time series key.
}
\description{
Read the release information of multiple time series vintages. The vintage range is given by the valid_from and valid_to parameter. By default, the entire release history is read.
}
\seealso{
Other time series collection functions: 
\code{\link{add_ts_to_collection}()},
\code{\link{create_collection}()},
\code{\link{delete_collection}()},
\code{\link{list_collections}()},
\code{\link{read_collection_keys}()},
\code{\link{read_collection_ts}()},
\code{\link{read_collection_ts_history}()},
\code{\link{read_collection_ts_metadata}()},
\code{\link{read_collection_ts_release}()},
\code{\link{read_collection_ts_release_future}()},
\code{\link{read_collection_ts_write_time}()},
\code{\link{remove_ts_from_collection}()}
}
\concept{time series collection functions}
