% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collections.R
\name{read_collection_ts_write_time}
\alias{read_collection_ts_write_time}
\title{Read collection time series vintage write time}
\usage{
read_collection_ts_write_time(
  collection,
  owner = "self",
  valid_on = Sys.Date(),
  ignore_missing = F
)
}
\arguments{
\item{collection}{Name of the time series collection}

\item{owner}{Username of the owner of the time series collection}

\item{valid_on}{Selects the time series vintage with the vintage date equal to or before this date}

\item{ignore_missing}{Whether to ignore missing or forbidden time series when requesting time series data}
}
\value{
Table with the vintage write time for every time series key.
}
\description{
Read the time at which a time series vintage was written. The vintage is specified by the valid_on parameter.
}
\seealso{
Other time series collection functions: 
\code{\link{add_ts_to_collection}()},
\code{\link{create_collection}()},
\code{\link{delete_collection}()},
\code{\link{list_collections}()},
\code{\link{read_collection_keys}()},
\code{\link{read_collection_ts}()},
\code{\link{read_collection_ts_history}()},
\code{\link{read_collection_ts_metadata}()},
\code{\link{read_collection_ts_release}()},
\code{\link{read_collection_ts_release_future}()},
\code{\link{read_collection_ts_release_history}()},
\code{\link{remove_ts_from_collection}()}
}
\concept{time series collection functions}
