% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{read_dataset_ts}
\alias{read_dataset_ts}
\title{Read dataset time series}
\usage{
read_dataset_ts(dataset, valid_on = Sys.Date(), ignore_missing = F)
}
\arguments{
\item{dataset}{Name of the dataset. A dataset is a group of time series with a common theme. Every time series can belong to only one dataset.}

\item{valid_on}{Selects the time series vintage with the vintage date equal to or before this date}

\item{ignore_missing}{Whether to ignore missing or forbidden time series when requesting time series data}
}
\value{
List of time series. Regular time series have the class ts and irregular time series the class xts.
}
\description{
Read the time series in a dataset. The time series vintage is specified by the valid_on parameter.
}
\seealso{
Other dataset functions: 
\code{\link{create_dataset}()},
\code{\link{delete_dataset}()},
\code{\link{list_datasets}()},
\code{\link{read_dataset_keys}()},
\code{\link{read_dataset_ts_history}()},
\code{\link{read_dataset_ts_metadata}()},
\code{\link{read_dataset_ts_release}()},
\code{\link{read_dataset_ts_release_future}()},
\code{\link{read_dataset_ts_release_history}()},
\code{\link{read_dataset_ts_write_time}()},
\code{\link{write_dataset_ts_release}()}
}
\concept{dataset functions}
