% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{read_dataset_ts_metadata}
\alias{read_dataset_ts_metadata}
\title{Read dataset time series metadata}
\usage{
read_dataset_ts_metadata(dataset, locale = NULL, ignore_missing = F)
}
\arguments{
\item{dataset}{Name of the dataset. A dataset is a group of time series with a common theme. Every time series can belong to only one dataset.}

\item{locale}{The locale of the metadata. Can be any string, but ISO codes are recommended (such as 'en', 'de', 'fr', 'it'). Set to NULL for unlocalized metadata (default).}

\item{ignore_missing}{Whether to ignore missing or forbidden time series when requesting time series data}
}
\value{
List of time series metadata. Each list element is named by the corresponding time series key and contains the metadata as a named list.
}
\description{
Read the time series metadata of a particular locale of the time series in a dataset.
}
\seealso{
Other dataset functions: 
\code{\link{create_dataset}()},
\code{\link{delete_dataset}()},
\code{\link{list_datasets}()},
\code{\link{read_dataset_keys}()},
\code{\link{read_dataset_ts}()},
\code{\link{read_dataset_ts_history}()},
\code{\link{read_dataset_ts_release}()},
\code{\link{read_dataset_ts_release_future}()},
\code{\link{read_dataset_ts_release_history}()},
\code{\link{read_dataset_ts_write_time}()},
\code{\link{write_dataset_ts_release}()}
}
\concept{dataset functions}
