% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{read_dataset_ts_release_future}
\alias{read_dataset_ts_release_future}
\title{Read dataset time series release future}
\usage{
read_dataset_ts_release_future(dataset, ignore_missing = F)
}
\arguments{
\item{dataset}{Name of the dataset. A dataset is a group of time series with a common theme. Every time series can belong to only one dataset.}

\item{ignore_missing}{Whether to ignore missing or forbidden time series when requesting time series data}
}
\value{
Table with release topic, year, period and time for every time series key and future release.
}
\description{
Read the release information of future time series vintages.
}
\seealso{
Other dataset functions: 
\code{\link{create_dataset}()},
\code{\link{delete_dataset}()},
\code{\link{list_datasets}()},
\code{\link{read_dataset_keys}()},
\code{\link{read_dataset_ts}()},
\code{\link{read_dataset_ts_history}()},
\code{\link{read_dataset_ts_metadata}()},
\code{\link{read_dataset_ts_release}()},
\code{\link{read_dataset_ts_release_history}()},
\code{\link{read_dataset_ts_write_time}()},
\code{\link{write_dataset_ts_release}()}
}
\concept{dataset functions}
