% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{read_dataset_ts_release_history}
\alias{read_dataset_ts_release_history}
\title{Read dataset time series release history}
\usage{
read_dataset_ts_release_history(
  dataset,
  valid_from = as.Date("1900-01-01"),
  valid_to = Sys.Date(),
  ignore_missing = F
)
}
\arguments{
\item{dataset}{Name of the dataset. A dataset is a group of time series with a common theme. Every time series can belong to only one dataset.}

\item{valid_from}{Start date of the vintage range}

\item{valid_to}{End date of the vintage range}

\item{ignore_missing}{Whether to ignore missing or forbidden time series when requesting time series data}
}
\value{
Table with vintage date, release topic, year, period and time for every time series key.
}
\description{
Read the release information of multiple time series vintages. The vintage range is given by the valid_from and valid_to parameter. By default, the entire release history is read.
}
\seealso{
Other dataset functions: 
\code{\link{create_dataset}()},
\code{\link{delete_dataset}()},
\code{\link{list_datasets}()},
\code{\link{read_dataset_keys}()},
\code{\link{read_dataset_ts}()},
\code{\link{read_dataset_ts_history}()},
\code{\link{read_dataset_ts_metadata}()},
\code{\link{read_dataset_ts_release}()},
\code{\link{read_dataset_ts_release_future}()},
\code{\link{read_dataset_ts_write_time}()},
\code{\link{write_dataset_ts_release}()}
}
\concept{dataset functions}
