% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{read_dataset_ts_write_time}
\alias{read_dataset_ts_write_time}
\title{Read dataset time series vintage write time}
\usage{
read_dataset_ts_write_time(dataset, valid_on = Sys.Date(), ignore_missing = F)
}
\arguments{
\item{dataset}{Name of the dataset. A dataset is a group of time series with a common theme. Every time series can belong to only one dataset.}

\item{valid_on}{Selects the time series vintage with the vintage date equal to or before this date}

\item{ignore_missing}{Whether to ignore missing or forbidden time series when requesting time series data}
}
\value{
Table with the vintage write time for every time series key.
}
\description{
Read the time at which a time series vintage was written. The vintage is specified by the valid_on parameter.
}
\seealso{
Other dataset functions: 
\code{\link{create_dataset}()},
\code{\link{delete_dataset}()},
\code{\link{list_datasets}()},
\code{\link{read_dataset_keys}()},
\code{\link{read_dataset_ts}()},
\code{\link{read_dataset_ts_history}()},
\code{\link{read_dataset_ts_metadata}()},
\code{\link{read_dataset_ts_release}()},
\code{\link{read_dataset_ts_release_future}()},
\code{\link{read_dataset_ts_release_history}()},
\code{\link{write_dataset_ts_release}()}
}
\concept{dataset functions}
