% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series.R
\name{read_ts}
\alias{read_ts}
\title{Read time series}
\usage{
read_ts(ts_keys, valid_on = Sys.Date(), ignore_missing = F)
}
\arguments{
\item{ts_keys}{Character vector of time series identifiers (keys)}

\item{valid_on}{Selects the time series vintage with the vintage date equal to or before this date}

\item{ignore_missing}{Whether to ignore missing or forbidden time series when requesting time series data}
}
\value{
List of time series. Regular time series have the class ts and irregular time series the class xts.
}
\description{
Read time series given by their unique identifiers (keys). The vintage is specified by the valid_on parameter.
By default, the most recent vintage is read.
}
\seealso{
Other time series functions: 
\code{\link{delete_ts}()},
\code{\link{find_ts}()},
\code{\link{read_ts_dataset}()},
\code{\link{read_ts_history}()},
\code{\link{read_ts_metadata}()},
\code{\link{read_ts_release}()},
\code{\link{read_ts_release_future}()},
\code{\link{read_ts_release_history}()},
\code{\link{read_ts_write_time}()},
\code{\link{rename_ts}()},
\code{\link{write_ts}()},
\code{\link{write_ts_dataset}()},
\code{\link{write_ts_metadata}()},
\code{\link{write_ts_release}()}
}
\concept{time series functions}
