% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series.R
\name{read_ts_release_future}
\alias{read_ts_release_future}
\title{Read time series release future}
\usage{
read_ts_release_future(ts_keys, ignore_missing = F)
}
\arguments{
\item{ts_keys}{Character vector of time series identifiers (keys)}

\item{ignore_missing}{Whether to ignore missing or forbidden time series when requesting time series data}
}
\value{
Table with release topic, year, period and time for every time series key and future release.
}
\description{
Read the release information of future time series vintages.
}
\seealso{
Other time series functions: 
\code{\link{delete_ts}()},
\code{\link{find_ts}()},
\code{\link{read_ts}()},
\code{\link{read_ts_dataset}()},
\code{\link{read_ts_history}()},
\code{\link{read_ts_metadata}()},
\code{\link{read_ts_release}()},
\code{\link{read_ts_release_history}()},
\code{\link{read_ts_write_time}()},
\code{\link{rename_ts}()},
\code{\link{write_ts}()},
\code{\link{write_ts_dataset}()},
\code{\link{write_ts_metadata}()},
\code{\link{write_ts_release}()}
}
\concept{time series functions}
