% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series.R
\name{read_ts_release_history}
\alias{read_ts_release_history}
\title{Read time series vintage release history}
\usage{
read_ts_release_history(
  ts_keys,
  valid_from = NULL,
  valid_to = NULL,
  ignore_missing = F
)
}
\arguments{
\item{ts_keys}{Character vector of time series identifiers (keys)}

\item{valid_from}{Start date of the vintage range}

\item{valid_to}{End date of the vintage range}

\item{ignore_missing}{Whether to ignore missing or forbidden time series when requesting time series data}
}
\value{
Table with vintage date, release topic, year, period and time for every time series key.
}
\description{
Read the release information of multiple time series vintages.
The vintage range is given by the valid_from and valid_to parameters.
By default, the entire release history is read.
}
\seealso{
Other time series functions: 
\code{\link{delete_ts}()},
\code{\link{find_ts}()},
\code{\link{read_ts}()},
\code{\link{read_ts_dataset}()},
\code{\link{read_ts_history}()},
\code{\link{read_ts_metadata}()},
\code{\link{read_ts_release}()},
\code{\link{read_ts_release_future}()},
\code{\link{read_ts_write_time}()},
\code{\link{rename_ts}()},
\code{\link{write_ts}()},
\code{\link{write_ts_dataset}()},
\code{\link{write_ts_metadata}()},
\code{\link{write_ts_release}()}
}
\concept{time series functions}
