% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series.R
\name{read_ts_write_time}
\alias{read_ts_write_time}
\title{Read time series vintage write time}
\usage{
read_ts_write_time(ts_keys, valid_on = Sys.Date(), ignore_missing = F)
}
\arguments{
\item{ts_keys}{Character vector of time series identifiers (keys)}

\item{valid_on}{Selects the time series vintage with the vintage date equal to or before this date}

\item{ignore_missing}{Whether to ignore missing or forbidden time series when requesting time series data}
}
\value{
Table with the vintage write time for every time series key.
}
\description{
Read the time at which a time series vintage was written.
The vintage is specified by the valid_on parameter.
}
\seealso{
Other time series functions: 
\code{\link{delete_ts}()},
\code{\link{find_ts}()},
\code{\link{read_ts}()},
\code{\link{read_ts_dataset}()},
\code{\link{read_ts_history}()},
\code{\link{read_ts_metadata}()},
\code{\link{read_ts_release}()},
\code{\link{read_ts_release_future}()},
\code{\link{read_ts_release_history}()},
\code{\link{rename_ts}()},
\code{\link{write_ts}()},
\code{\link{write_ts_dataset}()},
\code{\link{write_ts_metadata}()},
\code{\link{write_ts_release}()}
}
\concept{time series functions}
