% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{read_user_quota}
\alias{read_user_quota}
\title{Read user quota}
\usage{
read_user_quota(username = "self")
}
\arguments{
\item{username}{Username of the time series database user}
}
\value{
Table with quota information.
}
\description{
Check the number of time series downloads remaining in the current subscription year.
}
