% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_sets.R
\name{remove_ts_from_access_set}
\alias{remove_ts_from_access_set}
\title{Remove time series from access set}
\usage{
remove_ts_from_access_set(access_set, ts_keys)
}
\arguments{
\item{access_set}{Name of the access set. An access set is a group of time series for which a permission (read_quota, read, read_before_release, write) can be granted to individual users. Every time series can be a member of multiple access sets.}

\item{ts_keys}{Character vector of time series identifiers (keys)}
}
\description{
Remove time series (given by their keys) from a time series access set.
}
\seealso{
Other access set functions: 
\code{\link{add_ts_to_access_set}()},
\code{\link{create_access_set}()},
\code{\link{delete_access_set}()},
\code{\link{list_access_sets}()},
\code{\link{read_access_set_keys}()}
}
\concept{access set functions}
