% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collections.R
\name{remove_ts_from_collection}
\alias{remove_ts_from_collection}
\title{Remove time series from collection}
\usage{
remove_ts_from_collection(collection, owner = "self", ts_keys)
}
\arguments{
\item{collection}{Name of the time series collection}

\item{owner}{Username of the owner of the time series collection}

\item{ts_keys}{Character vector of time series identifiers (keys)}
}
\description{
Remove existing time series (given by their keys) from a time series collection.
}
\seealso{
Other time series collection functions: 
\code{\link{add_ts_to_collection}()},
\code{\link{create_collection}()},
\code{\link{delete_collection}()},
\code{\link{list_collections}()},
\code{\link{read_collection_keys}()},
\code{\link{read_collection_ts}()},
\code{\link{read_collection_ts_history}()},
\code{\link{read_collection_ts_metadata}()},
\code{\link{read_collection_ts_release}()},
\code{\link{read_collection_ts_release_future}()},
\code{\link{read_collection_ts_release_history}()},
\code{\link{read_collection_ts_write_time}()}
}
\concept{time series collection functions}
