% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{remove_user_access_sets}
\alias{remove_user_access_sets}
\title{Remove user access sets}
\usage{
remove_user_access_sets(username = "self", access_sets)
}
\arguments{
\item{username}{Username of the time series database user}

\item{access_sets}{Names of the access sets. An access set is a group of time series for which a permission (one of read_quota, read, read_before_release or write) can be granted to users. A time series can be a member of multiple access sets.}
}
\description{
Remove access sets and associated permissions from a user.
}
\seealso{
Other user management functions: 
\code{\link{add_user_access_sets}()},
\code{\link{create_user_api_key}()},
\code{\link{create_user_quota}()},
\code{\link{delete_user_quota}()},
\code{\link{list_user_access_sets}()},
\code{\link{list_users}()},
\code{\link{write_user_quota}()}
}
\concept{user management functions}
