% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series.R
\name{rename_ts}
\alias{rename_ts}
\title{Rename time series}
\usage{
rename_ts(ts_keys, ts_keys_new, ignore_missing = F)
}
\arguments{
\item{ts_keys}{Existing time series keys}

\item{ts_keys_new}{New time series keys}

\item{ignore_missing}{Whether to ignore missing or forbidden time series when requesting time series data}
}
\description{
Change the time series' unique identifiers (keys).
The length of the parameters ts_keys and ts_keys_new must be the same.
}
\seealso{
Other time series functions: 
\code{\link{delete_ts}()},
\code{\link{find_ts}()},
\code{\link{read_ts}()},
\code{\link{read_ts_dataset}()},
\code{\link{read_ts_history}()},
\code{\link{read_ts_metadata}()},
\code{\link{read_ts_release}()},
\code{\link{read_ts_release_future}()},
\code{\link{read_ts_release_history}()},
\code{\link{read_ts_write_time}()},
\code{\link{write_ts}()},
\code{\link{write_ts_dataset}()},
\code{\link{write_ts_metadata}()},
\code{\link{write_ts_release}()}
}
\concept{time series functions}
