% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_config}
\alias{set_config}
\title{Set package configuration}
\usage{
set_config(
  oauth_client_id = NULL,
  oauth_token_url = NULL,
  oauth_auth_url = NULL,
  oauth_auth_device_url = NULL,
  oauth_flow = NULL,
  oauth_redirect_url = NULL,
  api_key = NULL,
  url_staging = NULL,
  url_production = NULL,
  url_test = NULL,
  environment = NULL,
  access_type = NULL,
  read_before_release = NULL
)
}
\arguments{
\item{oauth_client_id}{OAuth client ID.}

\item{oauth_token_url}{OAuth token URL}

\item{oauth_auth_url}{OAuth authorization URL}

\item{oauth_auth_device_url}{OAuth device authorization URL}

\item{oauth_flow}{OAuth authorization flow. Must be either 'device' (default for hosted sessions) or 'code' (default otherwise).}

\item{oauth_redirect_url}{OAuth redirect URL}

\item{api_key}{API key for programmatic access. Use \code{\link[tsdbapi]{create_user_api_key}} to create an API key.}

\item{url_staging}{URL of staging API}

\item{url_production}{URL of production API}

\item{url_test}{URL of test API}

\item{environment}{Whether to use the production, staging or test API. Must be one of 'production', 'staging' or 'test.}

\item{access_type}{How to access time series data. Must be one of 'auth' (the default), 'public' or 'preview'.
The access types 'public' and 'preview' bypass authentication.
Use the access type 'public' to read public time series and the access type 'preview' to read time series previews (latest 2 years of data missing).
Use 'auth' for authenticated access.}

\item{read_before_release}{Whether to read time series vintages before their official release. Defaults to TRUE. This option will only have
an effect if you have pre release access to the requested time series.}
}
\description{
Set the package configuration options for the current session. Every option is initialized with the corresponding environment variable on package load.
}
\seealso{
Other package configuration functions: 
\code{\link{get_config}()}
}
\concept{package configuration functions}
