% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{write_dataset_ts_release}
\alias{write_dataset_ts_release}
\title{Write dataset time series release}
\usage{
write_dataset_ts_release(
  dataset,
  release_topic,
  release_year,
  release_period,
  valid_on = Sys.Date(),
  ignore_missing = F
)
}
\arguments{
\item{dataset}{Name of the dataset. A dataset is a group of time series with a common theme. Every time series can belong to only one dataset.}

\item{release_topic}{Topic of the release. For example 'kofbaro'.}

\item{release_year}{Year of the release}

\item{release_period}{Period of the release. For monthly releases a number from 1 to 12. For quarterly releases a number from 1 to 4. For half-yearly releases a number from 1 to 2. For yearly releases always 1.}

\item{valid_on}{Selects the time series vintage with the vintage date equal to or before this date}

\item{ignore_missing}{Whether to ignore missing or forbidden time series when requesting time series data}
}
\description{
Write the release information for a vintage of the time series in a dataset. The vintage is specified by the valid_on parameter.
}
\seealso{
Other dataset functions: 
\code{\link{create_dataset}()},
\code{\link{delete_dataset}()},
\code{\link{list_datasets}()},
\code{\link{read_dataset_keys}()},
\code{\link{read_dataset_ts}()},
\code{\link{read_dataset_ts_history}()},
\code{\link{read_dataset_ts_metadata}()},
\code{\link{read_dataset_ts_release}()},
\code{\link{read_dataset_ts_release_future}()},
\code{\link{read_dataset_ts_release_history}()},
\code{\link{read_dataset_ts_write_time}()}
}
\concept{dataset functions}
