% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series.R
\name{write_ts}
\alias{write_ts}
\title{Write time series}
\usage{
write_ts(
  ts_list,
  valid_from = Sys.Date(),
  access_sets = character(),
  release_topic = NULL,
  release_year = NULL,
  release_period = NULL,
  access = NULL,
  pre_release_access = NULL,
  release_date = NULL
)
}
\arguments{
\item{ts_list}{Named list of time series of class, zoo or xts with time series keys as list names}

\item{valid_from}{Vintage date of the time series. Defaults to the current date.}

\item{access_sets}{Names of the access sets the time series are added to.
If new time series are written, you must provide at least one access set name.}

\item{release_topic}{Topic of the release. For example 'kofbaro'.}

\item{release_year}{Year of the release}

\item{release_period}{Period of the release. For monthly releases a number from 1 to 12. For quarterly releases a number from 1 to 4. For half-yearly releases a number from 1 to 2. For yearly releases always 1.}

\item{access}{The access level of the time series. This parameter is obsolete once the Time Series DB API v1 is discontinued.}

\item{pre_release_access}{The access level of the time series before the time of release. This parameter is obsolete once the Time Series DB API v1 is discontinued.}

\item{release_date}{Time of the release (POSIXct). This parameter is obsolete once the Time Series DB API v1 is discontinued.}
}
\description{
Write time series as a particular vintage (given by the valid_from parameter).
}
\seealso{
Other time series functions: 
\code{\link{delete_ts}()},
\code{\link{find_ts}()},
\code{\link{read_ts}()},
\code{\link{read_ts_dataset}()},
\code{\link{read_ts_history}()},
\code{\link{read_ts_metadata}()},
\code{\link{read_ts_release}()},
\code{\link{read_ts_release_future}()},
\code{\link{read_ts_release_history}()},
\code{\link{read_ts_write_time}()},
\code{\link{rename_ts}()},
\code{\link{write_ts_dataset}()},
\code{\link{write_ts_metadata}()},
\code{\link{write_ts_release}()}
}
\concept{time series functions}
