% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series.R
\name{write_ts_dataset}
\alias{write_ts_dataset}
\title{Write time series dataset}
\usage{
write_ts_dataset(ts_keys, dataset)
}
\arguments{
\item{ts_keys}{Character vector of time series identifiers (keys)}

\item{dataset}{Name of the dataset. A dataset is a group of time series with a common theme. Every time series can belong to only one dataset.}
}
\description{
Assign time series to a dataset. Every time series can belong to only one dataset.
}
\seealso{
Other time series functions: 
\code{\link{delete_ts}()},
\code{\link{find_ts}()},
\code{\link{read_ts}()},
\code{\link{read_ts_dataset}()},
\code{\link{read_ts_history}()},
\code{\link{read_ts_metadata}()},
\code{\link{read_ts_release}()},
\code{\link{read_ts_release_future}()},
\code{\link{read_ts_release_history}()},
\code{\link{read_ts_write_time}()},
\code{\link{rename_ts}()},
\code{\link{write_ts}()},
\code{\link{write_ts_metadata}()},
\code{\link{write_ts_release}()}
}
\concept{time series functions}
