% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series.R
\name{write_ts_metadata}
\alias{write_ts_metadata}
\title{Write time series metadata}
\usage{
write_ts_metadata(metadata_list, locale = NULL, overwrite = F)
}
\arguments{
\item{metadata_list}{List of time series metadata named by the time series keys.
Each list element contains the metadata of a time series as a named list (key-value pairs).}

\item{locale}{The locale of the metadata. Can be any string, but ISO codes are recommended (such as 'en', 'de', 'fr', 'it'). Set to NULL for unlocalized metadata (default).}

\item{overwrite}{If TRUE, the existing metadata is replaced completely. If FALSE (default), new fields are added and existing fields are updated.}
}
\description{
Write time series metadata of a particular locale.
}
\seealso{
Other time series functions: 
\code{\link{delete_ts}()},
\code{\link{find_ts}()},
\code{\link{read_ts}()},
\code{\link{read_ts_dataset}()},
\code{\link{read_ts_history}()},
\code{\link{read_ts_metadata}()},
\code{\link{read_ts_release}()},
\code{\link{read_ts_release_future}()},
\code{\link{read_ts_release_history}()},
\code{\link{read_ts_write_time}()},
\code{\link{rename_ts}()},
\code{\link{write_ts}()},
\code{\link{write_ts_dataset}()},
\code{\link{write_ts_release}()}
}
\concept{time series functions}
