% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series.R
\name{write_ts_release}
\alias{write_ts_release}
\title{Write time series release}
\usage{
write_ts_release(
  ts_keys,
  release_topic,
  release_year,
  release_period,
  valid_on = Sys.Date()
)
}
\arguments{
\item{ts_keys}{Character vector of time series identifiers (keys)}

\item{release_topic}{Topic of the release. For example 'kofbaro'.}

\item{release_year}{Year of the release}

\item{release_period}{Period of the release. For monthly releases a number from 1 to 12. For quarterly releases a number from 1 to 4. For half-yearly releases a number from 1 to 2. For yearly releases always 1.}

\item{valid_on}{Selects the time series vintage with the vintage date equal to or before this date}
}
\description{
Write release information for a time series vintage. The vintage is specified by the valid_on parameter.
}
\seealso{
Other time series functions: 
\code{\link{delete_ts}()},
\code{\link{find_ts}()},
\code{\link{read_ts}()},
\code{\link{read_ts_dataset}()},
\code{\link{read_ts_history}()},
\code{\link{read_ts_metadata}()},
\code{\link{read_ts_release}()},
\code{\link{read_ts_release_future}()},
\code{\link{read_ts_release_history}()},
\code{\link{read_ts_write_time}()},
\code{\link{rename_ts}()},
\code{\link{write_ts}()},
\code{\link{write_ts_dataset}()},
\code{\link{write_ts_metadata}()}
}
\concept{time series functions}
