% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{write_user_quota}
\alias{write_user_quota}
\title{Write user quota}
\usage{
write_user_quota(
  username,
  current_year_quota = NULL,
  subscription_annual_quota = NULL
)
}
\arguments{
\item{username}{Username of the time series database user}

\item{current_year_quota}{Download quota of the current year. It is automatically set to subscription_annual_quota at the start of a subscription year.}

\item{subscription_annual_quota}{Annual download quota of data service subscription}
}
\description{
Set a user's quota for the current subscription year and/or the subscription's default annual quota.
}
\seealso{
Other user management functions: 
\code{\link{add_user_access_sets}()},
\code{\link{create_user_api_key}()},
\code{\link{create_user_quota}()},
\code{\link{delete_user_quota}()},
\code{\link{list_user_access_sets}()},
\code{\link{list_users}()},
\code{\link{remove_user_access_sets}()}
}
\concept{user management functions}
