% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series.R
\name{read_ts_history}
\alias{read_ts_history}
\title{Read time series history}
\usage{
read_ts_history(
  ts_keys,
  valid_from = NULL,
  valid_to = Sys.Date(),
  ignore_missing = F
)
}
\arguments{
\item{ts_keys}{Character vector of time series identifiers (keys)}

\item{valid_from}{Start date of the vintage range}

\item{valid_to}{End date of the vintage range}

\item{ignore_missing}{Whether to ignore missing or forbidden time series when requesting time series data}
}
\value{
List of time series.
Regular time series have the class ts and irregular time series the class xts.
The name of each time series is the time series key concatenated with the vintage date in format YYYYMMDD.
}
\description{
Read multiple time series vintages. The vintage range is given by the valid_from and valid_to parameters. By default, the entire history is read.
}
\seealso{
Other time series functions: 
\code{\link{delete_ts}()},
\code{\link{find_ts}()},
\code{\link{read_ts}()},
\code{\link{read_ts_dataset}()},
\code{\link{read_ts_metadata}()},
\code{\link{read_ts_release}()},
\code{\link{read_ts_release_future}()},
\code{\link{read_ts_release_history}()},
\code{\link{read_ts_write_time}()},
\code{\link{rename_ts}()},
\code{\link{write_ts}()},
\code{\link{write_ts_dataset}()},
\code{\link{write_ts_metadata}()},
\code{\link{write_ts_release}()}
}
\concept{time series functions}
